# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Tractive API
  version: 1.0.0
  description: API documentation for Tractive backend
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /api/transporters/fleet/{id}/payments:
    get:
      summary: List fleet payments for a fleet
      description: Buyer sees their own fleet payments for the fleet. Fleet owner transporter sees all payments for that fleet. Admin can also view all payments.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Fleet payments list
          content:
            application/json:
              example:
                success: true
                data:
                  - _id: "67defleetpay1234567890abc"
                    amount: 210000
                    paymentMethod: "bank_transfer"
                    status: "pending"
                    buyer:
                      _id: "67debuyer1234567890abc0"
                      name: "Buyer One"
                      email: "buyer@example.com"
                    transporter:
                      _id: "67detrans1234567890abc1"
                      businessName: "North Haul Logistics Ltd"
                    fleetBid:
                      _id: "67defleetbid1234567890abc"
                      amount: 180000
                      counterAmount: 210000
                      status: "accepted"
    post:
      summary: Create fleet payment
      description: Buyer creates a fleet payment record from an accepted fleet bid or the fleet listed price. This supports manual approval flows like cash or bank transfer. The endpoint also creates or reuses a linked fleet booking in `pending_payment` state.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fleetBidId:
                  type: string
                amount:
                  type: number
                paymentMethod:
                  type: string
                  enum: [cash, bank_transfer, card]
                note:
                  type: string
            example:
              fleetBidId: "67defleetbid1234567890abc"
              amount: 210000
              paymentMethod: "bank_transfer"
              note: "Paid via bank transfer for Kaduna to Lagos route"
      responses:
        '201':
          description: Fleet payment created
          content:
            application/json:
              example:
                success: true
                data:
                  _id: "67defleetpay1234567890abc"
                  amount: 210000
                  paymentMethod: "bank_transfer"
                  status: "pending"
                  fleetBid: "67defleetbid1234567890abc"
                  booking:
                    _id: "67defleetbooking1234567890ab"
                    status: "pending_payment"
...
  /api/transporters/fleet/{id}/bookings:
    get:
      summary: List bookings for a fleet
      description: Buyer sees their own bookings for the fleet. Fleet owner transporter sees all bookings for that fleet. Admin can also view all bookings.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [pending_payment, confirmed, rejected, cancelled, completed]
      responses:
        '200':
          description: Fleet bookings for fleet
...
  /api/transporters/orders/{orderId}/status:
    patch:
      summary: Update transport status for assigned order
      description: Allows the assigned transporter or a seller/agent attached to any product in the order to update delivery progress and create a tracking event.
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                transportStatus:
                  type: string
                  enum: [pending, picked, on_transit, delivered]
            example:
              transportStatus: "on_transit"
      responses:
        '200':
          description: Updated order
...